requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.0
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
ffmpeg-python==0.2.0
boto3==1.34.34
python-multipart==0.0.9
imagehash==4.3.1
Pillow==10.2.0
numpy==1.26.3
tenacity==8.2.3
python-dotenv==1.0.1
